LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/tentative/shadowrootadoptedstylesheets/shadowrootadoptedstylesheets-async-fetch-mutation-reinsert-swap-expected.txt


FAIL Placeholders reinserted in swapped positions reflect new cascade order after fetch. assert_equals: Before swap: expected 2 placeholder(s). expected 2 but got 0
